LayoutTests/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-location-worker-expected.txt

CONSOLE MESSAGE: Cross-origin redirection to invalidurl: denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to data:,data%20url denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to invalidurl: denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to data:,data%20url denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to invalidurl: denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to data:,data%20url denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to invalidurl: denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to data:,data%20url denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to invalidurl: denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
CONSOLE MESSAGE: Cross-origin redirection to data:,data%20url denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.

PASS Redirect 301 in "follow" mode without location 
FAIL Redirect 301 in "manual" mode without location assert_equals: Response's status is 0 expected 0 but got 301
PASS Redirect 301 in "follow" mode with invalid location 
PASS Redirect 301 in "manual" mode with invalid location 
PASS Redirect 301 in "follow" mode with data location 
PASS Redirect 301 in "manual" mode with data location 
PASS Redirect 302 in "follow" mode without location 
FAIL Redirect 302 in "manual" mode without location assert_equals: Response's status is 0 expected 0 but got 302
PASS Redirect 302 in "follow" mode with invalid location 
PASS Redirect 302 in "manual" mode with invalid location 
PASS Redirect 302 in "follow" mode with data location 
PASS Redirect 302 in "manual" mode with data location 
PASS Redirect 303 in "follow" mode without location 
FAIL Redirect 303 in "manual" mode without location assert_equals: Response's status is 0 expected 0 but got 303
PASS Redirect 303 in "follow" mode with invalid location 
PASS Redirect 303 in "manual" mode with invalid location 
PASS Redirect 303 in "follow" mode with data location 
PASS Redirect 303 in "manual" mode with data location 
PASS Redirect 307 in "follow" mode without location 
FAIL Redirect 307 in "manual" mode without location assert_equals: Response's status is 0 expected 0 but got 307
PASS Redirect 307 in "follow" mode with invalid location 
PASS Redirect 307 in "manual" mode with invalid location 
PASS Redirect 307 in "follow" mode with data location 
PASS Redirect 307 in "manual" mode with data location 
PASS Redirect 308 in "follow" mode without location 
FAIL Redirect 308 in "manual" mode without location assert_equals: Response's status is 0 expected 0 but got 308
PASS Redirect 308 in "follow" mode with invalid location 
PASS Redirect 308 in "manual" mode with invalid location 
PASS Redirect 308 in "follow" mode with data location 
PASS Redirect 308 in "manual" mode with data location 
